backend/requirements.txt
# Requires Python >=3.11 for enum.StrEnum.
fastapi>=0.110,<1
httpx>=0.27,<1
pydantic>=2.6,<3
pytest>=8,<9
uvicorn[standard]>=0.27,<1
python-dotenv>=1,<2
langchain-deepseek>=1.0.1,<2
